requirements.txt
fastapi
uvicorn
torch
scikit-learn
pydantic
requests
python-dotenv
